+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Changes relevant to the users of python-tcod are documented here.
 This project adheres to [Semantic Versioning](https://semver.org/) since version `2.0.0`.
 
 ## [Unreleased]
+### Deprecated
+- Deprecated using `tcod` as an implicit alias for `libtcodpy`.
+  You should use `from tcod import libtcodpy` if you want to access this module.
+- Deprecated constants being held in `tcod`, get these from `tcod.constants` instead.
+- Deprecated `tcod.Console` which should be accessed from `tcod.console.Console` instead.
 
 ## [16.0.2] - 2023-06-02
 ### Fixed